move all the go files to the server/ folder
