+++ server/handlers/mocks.go
@@ -5,9 +5,9 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/Thiht/smocker/services"
-	"github.com/Thiht/smocker/templates"
-	"github.com/Thiht/smocker/types"
+	"github.com/Thiht/smocker/server/services"
+	"github.com/Thiht/smocker/server/templates"
+	"github.com/Thiht/smocker/server/types"
 	"github.com/labstack/echo"
 	log "github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v3"
