+++ server/middlewares.go
@@ -13,8 +13,8 @@ import (
 	"runtime"
 	"time"
 
-	"github.com/Thiht/smocker/services"
-	"github.com/Thiht/smocker/types"
+	"github.com/Thiht/smocker/server/services"
+	"github.com/Thiht/smocker/server/types"
 	"github.com/labstack/echo"
 	log "github.com/sirupsen/logrus"
 )
